services/node-engine/requirements.txt
accelerate==1.11.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
autopep8==2.3.2
billiard==4.2.2
bitsandbytes==0.48.2
blinker==1.9.0
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
fastapi==0.121.1
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
honcho==2.0.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
prompt_toolkit==3.0.52
psutil==7.1.3
pycodestyle==2.14.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
websocket-client==1.8.0
python-dotenv
pytest==8.2.2
httpx==0.27.0
grpcio
protobuf
grpcio-tools
